+++ UndertaleModLib/UndertaleDebugChunks.cs
@@ -32,7 +32,7 @@ namespace UndertaleModLib
         internal override void UnserializeChunk(UndertaleReader reader)
         {
             Chunks.Clear();
-            uint startPos = reader.Position;
+            long startPos = reader.Position;
             while (reader.Position < startPos + Length)
             {
                 UndertaleChunk chunk = reader.ReadUndertaleChunk();
